do not lock pod list during other operation
- pod list lock is dedicate for pod list internal access
- remove GetPod/RunPod legacy interface
- set pod return status success only if the pod return with all 0 value, i.e. set failed if user stop it.
- minor adjust of the pod tty lock
- move the demon methods on pod to a new file
- adjust the finish/shutdown processing
